+++ tests/pyccel/test_pyccel.py
@@ -803,6 +803,7 @@ def test_basic_header():
                                         "scripts/classes/classes_4.py",
                                         "scripts/classes/classes_5.py",
                                         "scripts/classes/classes_6.py",
+                                        "scripts/classes/classes_7.py",
                                         ] )
 @pytest.mark.parametrize( 'language', (
         pytest.param("python", marks = pytest.mark.python),